consumer_service/requirements.txt
pymongo==4.6.1
kafka-python